Fix potential credential recovery vulnerability
This replaces the insecure AES-CTR encryption that always used the same
IV/counter with AES-GCM-SIV with a random nonce for each encrypted value.
We now also derive an individual encryption key using HKDF to further
separate the values.

The old scheme is still supported when reading data, in particular,
to run migrations that convert existing data.