Fix server-side configuration of EAP authentication
The authentication objects were incorrect and assigned reversed.  For
EAP-TLS we keep that scheme, though, as it makes the distinction between
EAP-only and pubkey auth for the server easier.  The combined certificate+EAP
client authentication was also not handled correctly.

This also adds the ability to configure an EAP-Identity and fixes some
things in the GUI.